Add a marker to skip database migration tests older than 30 days when in fast mode (PR #4879)
Pull request opened by the merge tool on behalf of #4879
