+++ changelogs/unreleased/add-test-marker-migration.yml
@@ -1,0 +1,3 @@
+change-type: patch
+description: Add a marker to skip database migration tests older than 30 days when in fast mode
+destination-branches: [master, iso5]
