+++ tests/db/migration_tests/test_v3_to_v4.py
@@ -29,7 +29,6 @@ from inmanta.server.bootloader import InmantaBootloader
 
 
 @pytest.fixture
-@pytest.mark.slowtest
 async def migrate_v3_to_v4(hard_clean_db, hard_clean_db_post, postgresql_client: Connection, async_finalizer, server_config):
     # Get old tables
     with open(os.path.join(os.path.dirname(__file__), "dumps/v3.sql"), "r") as fh:
@@ -51,6 +50,7 @@ async def migrate_v3_to_v4(hard_clean_db, hard_clean_db_post, postgresql_client:
     await ibl.stop(timeout=15)
 
 
+@pytest.mark.slowtest
 async def test_db_migration(migrate_v3_to_v4, postgresql_client: Connection):
     for table_name in ["form", "formrecord", "resourceversionid"]:
         assert not await does_table_exist(postgresql_client, table_name)
