+++ tests/db/migration_tests/test_v4_to_v5.py
@@ -43,6 +43,7 @@ async def migrate_v4_to_v5(
     await ibl.stop(timeout=15)
 
 
+@pytest.mark.slowtest
 async def test_db_migration_compile_data(migrate_v4_to_v5, postgresql_client: Connection) -> None:
     compiles = await postgresql_client.fetch("SELECT * FROM public.compile;")
     for c in compiles:
@@ -52,6 +53,7 @@ async def test_db_migration_compile_data(migrate_v4_to_v5, postgresql_client: Co
         assert c["compile_data"] is None
 
 
+@pytest.mark.slowtest
 async def test_db_migration_environment_halt(migrate_v4_to_v5, postgresql_client: Connection) -> None:
     environments = await postgresql_client.fetch("SELECT * FROM public.environment;")
     for env in environments:
