+++ tests/db/migration_tests/test_v7_to_v17.py
@@ -30,7 +30,6 @@ from utils import retry_limited
 
 
 @pytest.fixture
-@pytest.mark.slowtest
 async def migrate_v7_to_v17(
     hard_clean_db, hard_clean_db_post, postgresql_client: Connection, server_config
 ) -> AsyncIterator[None]:
@@ -49,6 +48,7 @@ async def migrate_v7_to_v17(
     await ibl.stop(timeout=15)
 
 
+@pytest.mark.slowtest
 async def test_foreign_key_agent_to_agentinstance(migrate_v7_to_v17: None, postgresql_client: Connection) -> None:
     """
     Deleting an entry in the agentInstance table should not be allowed when it's references from the agent stable.
